requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
enum34==1.1.6
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipaddress==1.0.22
lxml==4.2.2
MySQL-python==1.2.5
numpy==1.14.5
parsel==1.4.0
Pillow==5.1.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
pytesseract==0.2.2
queuelib==1.5.0
requests==2.19.1
Scrapy==1.3.3
selenium==3.13.0
service-identity==17.0.0
six==1.11.0
tesseract==0.1.3
Twisted==18.4.0
urllib3==1.23
w3lib==1.19.0
zope.interface==4.5.0
